Move loot crate data sets to a new folder
