+++ apworld/brotato/test/data_sets/loot_crates.py
@@ -12,7 +12,7 @@ reason it couldn't be expanded to handle more in the future.
 from dataclasses import asdict, dataclass
 from typing import Any, Dict, List, Optional, Tuple, Union
 
-from ..constants import (
+from ...constants import (
     BASE_GAME_CHARACTERS,
     MAX_LEGENDARY_CRATE_DROP_GROUPS,
     MAX_LEGENDARY_CRATE_DROPS,
@@ -22,7 +22,7 @@ from ..constants import (
 
 
 @dataclass(frozen=True)
-class BrotatoTestOptions:
+class BrotatoLootCrateTestOptions:
     """Subset of the full options that we want to control for the test, with defaults.
 
     This avoids needing to specify all the options for the dataclass, and makes using it in the tests slightly more
@@ -37,7 +37,7 @@ class BrotatoTestOptions:
 
 
 @dataclass(frozen=True)
-class BrotatoTestExpectedResults:
+class BrotatoLootCrateTestExpectedResults:
     # An int value means all regions have the same number of crates.
     # A tuple of ints means region "Crate Group {i}" has number of crates in index [i]
     num_common_crate_regions: int
@@ -86,9 +86,9 @@ class BrotatoTestExpectedResults:
 
 
 @dataclass(frozen=True)
-class BrotatoTestDataSet:
-    options: BrotatoTestOptions
-    expected_results: BrotatoTestExpectedResults
+class BrotatoLootCrateTestDataSet:
+    options: BrotatoLootCrateTestOptions
+    expected_results: BrotatoLootCrateTestExpectedResults
     description: Optional[str] = None
 
     def test_name(self) -> str:
@@ -113,16 +113,16 @@ class BrotatoTestDataSet:
         return asdict(self.options)
 
 
-TEST_DATA_SETS: List[BrotatoTestDataSet] = [
-    BrotatoTestDataSet(
+TEST_DATA_SETS: List[BrotatoLootCrateTestDataSet] = [
+    BrotatoLootCrateTestDataSet(
         description="Easily divisible, common and legendary same (25 crates)",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=25,
             num_common_crate_drop_groups=5,
             num_legendary_crate_drops=25,
             num_legendary_crate_drop_groups=5,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=5,
             common_crates_per_region=5,
             num_legendary_crate_regions=5,
@@ -131,15 +131,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0, 6, 12, 18, 24),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Easily divisible, common and legendary same (30 crates)",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=30,
             num_common_crate_drop_groups=6,
             num_legendary_crate_drops=30,
             num_legendary_crate_drop_groups=6,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=6,
             common_crates_per_region=5,
             num_legendary_crate_regions=6,
@@ -148,15 +148,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0, 5, 10, 15, 20, 25),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Easily divisible, common and legendary are different",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=20,
             num_common_crate_drop_groups=2,
             num_legendary_crate_drops=30,
             num_legendary_crate_drop_groups=6,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=2,
             common_crates_per_region=10,
             num_legendary_crate_regions=6,
@@ -165,15 +165,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0, 5, 10, 15, 20, 25),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Unequal groups",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=16,
             num_common_crate_drop_groups=3,
             num_legendary_crate_drops=16,
             num_legendary_crate_drop_groups=3,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=3,
             common_crates_per_region=(6, 5, 5),
             num_legendary_crate_regions=3,
@@ -182,15 +182,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0, 10, 20),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Unequal groups, common and legendary are different",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=35,
             num_common_crate_drop_groups=15,
             num_legendary_crate_drops=25,
             num_legendary_crate_drop_groups=5,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             # Five "3's" and ten "2's", because the drops don't evenly divide into the groups
             num_common_crate_regions=15,
             common_crates_per_region=tuple(([3] * 5) + ([2] * 10)),
@@ -216,15 +216,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0, 6, 12, 18, 24),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Max possible groups and crates, more groups than req. wins, no DLC",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=MAX_NORMAL_CRATE_DROPS,
             num_common_crate_drop_groups=MAX_NORMAL_CRATE_DROP_GROUPS,
             num_legendary_crate_drops=MAX_LEGENDARY_CRATE_DROPS,
             num_legendary_crate_drop_groups=MAX_LEGENDARY_CRATE_DROP_GROUPS,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             # The number of groups will be set to 30 (default # of wins) when generated.
             num_common_crate_regions=30,
             common_crates_per_region=tuple(([2] * 20) + ([1] * 10)),
@@ -235,9 +235,9 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=tuple(range(30)),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Max wins, one crate per character, one group per character, no DLC",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_victories=BASE_GAME_CHARACTERS.num_characters,
             num_common_crate_drops=BASE_GAME_CHARACTERS.num_characters,
             # Assign one group per character, so each win makes more crates accessible.
@@ -245,7 +245,7 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             num_legendary_crate_drops=BASE_GAME_CHARACTERS.num_characters,
             num_legendary_crate_drop_groups=BASE_GAME_CHARACTERS.num_characters,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=BASE_GAME_CHARACTERS.num_characters,
             common_crates_per_region=tuple([1] * BASE_GAME_CHARACTERS.num_characters),
             num_legendary_crate_regions=BASE_GAME_CHARACTERS.num_characters,
@@ -255,16 +255,16 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=tuple(range(BASE_GAME_CHARACTERS.num_characters)),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Max number of crates, one group",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_victories=BASE_GAME_CHARACTERS.num_characters,
             num_common_crate_drops=MAX_NORMAL_CRATE_DROPS,
             num_common_crate_drop_groups=1,
             num_legendary_crate_drops=MAX_LEGENDARY_CRATE_DROPS,
             num_legendary_crate_drop_groups=1,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=1,
             common_crates_per_region=50,
             num_legendary_crate_regions=1,
@@ -274,15 +274,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0,),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="1 crate and 1 group",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=1,
             num_common_crate_drop_groups=1,
             num_legendary_crate_drops=1,
             num_legendary_crate_drop_groups=1,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=1,
             common_crates_per_region=1,
             num_legendary_crate_regions=1,
@@ -291,15 +291,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0,),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="2 crates, 1 group",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=2,
             num_common_crate_drop_groups=1,
             num_legendary_crate_drops=2,
             num_legendary_crate_drop_groups=1,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=1,
             common_crates_per_region=2,
             num_legendary_crate_regions=1,
@@ -308,15 +308,15 @@ TEST_DATA_SETS: List[BrotatoTestDataSet] = [
             wins_required_per_legendary_region=(0,),
         ),
     ),
-    BrotatoTestDataSet(
+    BrotatoLootCrateTestDataSet(
         description="Max number of crates, 1 common group, 2 legendary groups",
-        options=BrotatoTestOptions(
+        options=BrotatoLootCrateTestOptions(
             num_common_crate_drops=MAX_LEGENDARY_CRATE_DROPS,
             num_common_crate_drop_groups=1,
             num_legendary_crate_drops=MAX_LEGENDARY_CRATE_DROPS,
             num_legendary_crate_drop_groups=2,
         ),
-        expected_results=BrotatoTestExpectedResults(
+        expected_results=BrotatoLootCrateTestExpectedResults(
             num_common_crate_regions=1,
             common_crates_per_region=50,
             num_legendary_crate_regions=2,
